requirements.txt
aiogram>=3.2.0
python-dotenv>=1.0.0
google-generativeai>=0.3.1
Pillow>=10.0.0
fluent.runtime>=0.4